+++ silkworm/node/stagedsync/stages/stage_interhashes.cpp
@@ -24,6 +24,7 @@
 
 #include <silkworm/core/common/endian.hpp>
 #include <silkworm/core/common/lru_cache.hpp>
+#include <silkworm/core/execution/address.hpp>
 #include <silkworm/core/trie/nibbles.hpp>
 #include <silkworm/infra/common/decoding_exception.hpp>
 #include <silkworm/infra/common/stopwatch.hpp>
