+++ silkworm/node/stagedsync/stages/stage_interhashes/_test.cpp
@@ -16,6 +16,7 @@
 
 #include <catch2/catch.hpp>
 
+#include <silkworm/core/execution/address.hpp>
 #include <silkworm/core/trie/hash_builder.hpp>
 #include <silkworm/core/trie/nibbles.hpp>
 #include <silkworm/core/types/account.hpp>
