+++ silkworm/silkrpc/json/call.cpp
@@ -20,6 +20,7 @@
 #include <utility>
 
 #include <silkworm/core/common/util.hpp>
+#include <silkworm/core/execution/address.hpp>
 #include <silkworm/silkrpc/common/util.hpp>
 #include <silkworm/silkrpc/json/types.hpp>
