+++ silkworm/silkrpc/json/log.cpp
@@ -21,6 +21,7 @@
 #include <utility>
 
 #include <silkworm/core/common/util.hpp>
+#include <silkworm/core/execution/address.hpp>
 #include <silkworm/silkrpc/common/util.hpp>
 
 #include "types.hpp"
